requirements.txt
langchain_community
langchain-openai
langchain
openai
chainlit
chromadb
pypdf
python-dotenv
chromadb